+++ src-tauri/Cargo.toml
@@ -6,25 +6,23 @@ authors = ["hrzlgnm"]
 edition = "2021"
 
 [build-dependencies]
-tauri-build = { version = "1.5.5", features = [] }
+tauri-build = { version = "2.0.0-rc.0", features = [] }
 
 [dependencies]
-clap = { version = "4.5.17", features = ["derive"] }
-log = "0.4.22"
-mdns-sd = { version = "0.11.4", features = ["async", "log"] }
-serde = { version = "1.0.210", features = ["derive"] }
-serde_json = "1.0.128"
-tauri = { version = "1.8.0", features = [
-    "dialog-ask",
-    "http-api",
-    "http-request",
-    "process-exit",
-    "process-relaunch",
-    "shell-open",
-    "updater",
-    "window-all",
+mdns-sd = { git = "https://github.com/keepsimple1/mdns-sd.git", rev = "1e9840f5", default-features = false, features = [
+    "async",
+    "log",
 ] }
-tauri-plugin-log = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
+tauri = { version = "2.0.0-rc.0", features = [] }
+serde = { version = "1.0.204", features = ["derive"] }
+serde_json = "1.0.122"
+tauri-plugin-log = "2.0.0-rc.0"
+log = "0.4.22"
+tauri-plugin-shell = "2.0.0-rc.0"
+
+[lib]
+name = "mdns_browser_lib"
+crate-type = ["staticlib", "cdylib", "lib"]
 
 [features]
 # This feature is used for production builds or when a dev server is not specified, DO NOT REMOVE!!
