+++ src-tauri/capabilities/migrated.json
@@ -1,0 +1,12 @@
+{
+    "identifier": "migrated",
+    "description": "permissions that were migrated from v1",
+    "local": true,
+    "windows": [
+        "main",
+        "splashscreen"
+    ],
+    "permissions": [
+        "core:default"
+    ]
+}
